requirements.txt
langchain-core>=0.3.59,<0.4.0
langchain-openai>=0.2.14
langchain-community>=0.3.13
langchain-anthropic>=0.3.13
langchain-google-vertexai>=2.0.23
langchain-chroma>=0.1.0
google-cloud-aiplatform>=1.92.0
ragas>=0.2.9
boto3==1.34.0
chromadb==0.4.18
numpy<2.0.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv==1.0.1
tiktoken>=0.7,<1
beautifulsoup4==4.12.3
markdown==3.5.2
pytest==8.0.2
evaluate==0.4.1
datasets==2.18.0
pytest>=7.0.0
pytest-cov>=4.0.0